lib.rs: return values from read via a channel
Values are returned from FASTER via a callback into Rust which sends
the value along a channel. The callee is given the receiving end of the
channel.

In case the value is not found, the sender will be dropped so receiver
will throw error when trying to access value. It is the callee's
responsibility to handle the error as appropriate.